requirements.txt
django
djangorestframework
pandas
scikit-learn
xgboost
joblib
django-cors-headers
tensorflow
gunicorn
Pillow
numpy
